// src/modules/company/types/company-stream.types.ts
export interface CompanyDataStream {
  image: string;
  business_name: string;
  code: string;
  address: string;
  created_at: number;
  total_employee: string;
  updated_at: number;
  phone: string;
  province_id: number;
  name: string;
  tagline: string;
  id: string;
  district_id: number;
  job: string;
  village: string;
  table: string;
  prepared_process: boolean;
  city_id: number;
  status: boolean;
}